+++ psychopy/demos/coder/iohub/iosync/retracesynctiming.py
@@ -117,7 +117,7 @@ ai0_times=[]
 try:
     psychopy_mon_name='testMonitor'
     exp_code='events'
-    sess_code='S_{0}'.format(long(time.mktime(time.localtime())))
+    sess_code='S_{0}'.format(int(time.mktime(time.localtime())))
     iohub_config={
     "psychopy_monitor_name":psychopy_mon_name,
     "mcu.iosync.MCU":dict(serial_port='auto',monitor_event_types=['AnalogInputEvent']),#['DigitalInputEvent']),
